web/requirements.txt
pygdrive3
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
oauth2client
flask
pytz
pygdrive3
pyngrok
flask_ngrok
#pip install --upgrade oauth2client
google-cloud-pubsub
google-cloud